% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{fake_dem}
\alias{fake_dem}
\title{Get a fake DEM}
\usage{
fake_dem(r = NULL, true_dem = NULL, n_random_data = ncell(r) * 0.01,
  z_range = c(0, 500))
}
\arguments{
\item{r}{\code{\linkS4class{RasterLayer}}}

\item{true_dem}{\code{\linkS4class{RasterLayer}}}

\item{z_range}{}
}
\value{
\code{\linkS4class{RasterLayer}}
}
\description{
Get a fake DEM
}
\examples{
# gerenate the frame for the fake DEM
r <- raster(ncol = 200, nrow = 100)
projection(r) <- NA
extent(r) <- extent(0, 200, 0, 100)

# load a DEM to use as the true DEM
path <- system.file("external/dtm.grd", package="rdems")
dtm <- raster(path)

fdem <- fake_dem(r, dtm)
plot(fdem)
}
