% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{fake_voids}
\alias{fake_voids}
\title{Get a fake mask of voids}
\usage{
fake_voids(center, void_size, y, is_circular = FALSE)
}
\arguments{
\item{center}{dataframe or matrix. Coordinates.}

\item{void_size}{numeric}

\item{is_circular}{logical.}

\item{...}{Arguments passed on to \code{raster::rasterize}
\describe{
  \item{y}{Raster* object}
}}
}
\value{
\code{\linkS4class{RasterLayer}}
}
\description{
Get a fake mask of voids
}
\examples{
fdem <- fake_dem(r, dtm)
plot(fdem)

center <- sampleRandom(fdem, 10, xy = TRUE, sp = TRUE)
voids_mask <- fake_voids(center, 20, fdem)
plot(voids_mask)
}
