% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_voids_g06.R
\name{fill_voids_g06}
\alias{fill_voids_g06}
\title{Voids are filling by fitting the fill source to the elevations
  surrounded the void}
\usage{
fill_voids_g06(x, np = 3, show_progress_bar = TRUE)
}
\arguments{
\item{x}{\code{\linkS4class{Raster}}. The dem with voids should be the first
layer, the filling source or sources should be the other layer of layers.}

\item{np}{
degree of polynomial surface
}

\item{show_progress_bar}{logical.}
}
\value{
\code{\linkS4class{RasterStack}}. First layer has the dem, second
  layer is a metadata layer that indicate which fill source was uses in each
  void.
}
\description{
\code{fitFillSource} returns a DEM with its voids filled with a
  source locally fitted with \code{\link{fit_trend_surface}}.
}
\details{
If the length of the shorter side of the bounding box of the void if
  less than \code{10}, the void is not filled.

  To estimate the trend surface between DEM and fill source, the area of the
  bounding box is quadruplicate by side duplication (the void remain
  centered). The trend surface is used to correct the differences between DEM
  and fill source. If there are more than one filled source available for the
  void (i.e, \code{x} has more than two layers and them have data for the
  given void), the algorithm uses the median of the corrected difference to
  estimate which is the better choice to filling the void. To obtain the
  median, the algorithm only uses the pixels surrounding the void.
}
\examples{
r <- raster(ncol = 200, nrow = 100)
extent(r) <- extent(0, 200, 0, 100)
projection(r) <- projection(r) <- CRS("+init=epsg:32718")

###reference
set.seed(11)
reference <- fake_dem(r, n_random_data = 60, z_range = c(0, 600))
plot(reference)

set.seed(18)
fillSource <- fake_dem(r, z_range = c(0, 20))
fillSource <- fillSource - mean(fillSource[])
fillSource <- fillSource + reference
fillSource <- smooth_dem(fillSource, theta = 6)
plot(fillSource)

upperHalfCells <- 1:(ncell(r)/2)
fillSource1 <- fillSource
fillSource1[upperHalfCells] <- (fillSource * 2)[upperHalfCells]
fillSource2 <- fillSource * 2
fillSource2[upperHalfCells] <- fillSource[upperHalfCells]

p <- sampleRegular(reference, 10, sp = TRUE)
voidsMask <- fake_voids(p, void_size = 20, reference)

dem <- reference
dem[voidsMask] <- NA
plot(dem)

demF1 <- fill_voids_g06(stack(dem, fillSource1, fillSource2), show_progress_bar = TRUE)
plot(subset(demF1, 1))
plot(subset(demF1, 2))

plot(subset(demF1, 1) - reference)

hs <-
  hillShade(terrain(subset(demF1, 1), "slope"),
            terrain(subset(demF1, 1), "aspect"))
plot(hs , col = grey((0:255) / 255), legend = FALSE)
}
\references{
Grohman, G., Kroenung, G., Strebeck, J., 2006. Filling SRTM
  voids: the delta surface fill method. Photogramm. Eng. Remote Sensing 72,
  213-216.
}
