% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_voids_k08.R
\name{fill_voids_k08}
\alias{fill_voids_k08}
\title{Void filling using derivatives}
\usage{
fill_voids_k08(x, slp, asp, show_progress_bar = TRUE)
}
\arguments{
\item{x}{\code{\linkS4class{RasterLayer}}. DEM with voids.It must be in a
cartographic projection with the same unit that the elevation data (meters
is strongly recommended).}

\item{slp}{\code{\linkS4class{RasterLayer}}. Slope in radians derivate from the fill source.}

\item{asp}{\code{\linkS4class{RasterLayer}}. Aspect in radians derivate from the fill source.}

\item{show_progress_bar}{logical.}
}
\description{
Void filling using trigonometric, elevation of pixels
  surrounding the voids, and slope and aspect from the fill source.
}
\details{
This function implement the algorithm developed by Karkee et al.
  (2008) (see references). The filling process iteratively grows from the
  surrounding pixels to the center of the void.
}
\examples{
set.seed(11)
reference <- fake_dem(n_random_data = 60, z_range = c(0, 600))
plot(reference)

set.seed(18)
fillSource <- fake_dem(z_range = c(0, 20))
fillSource <- fillSource - mean(fillSource[])
fillSource <- fillSource + reference
fillSource <- smooth_dem(fillSource, theta = 6)
plot(fillSource)

p <- sampleRegular(reference, 10, sp = TRUE)
voidsMask <- fake_voids(p, void_size = 20, reference)

dem <- reference
dem[voidsMask] <- NA
plot(dem)

slp <- terrain(fillSource, "slope")
asp <- terrain(fillSource, "aspect")
demF2 <- fill_voids_k08(dem, slp, asp, show_progress_bar = TRUE)

plot(demF2)

plot(demF2 - reference)

hs <- hillShade(terrain(demF2, "slope"), terrain(demF2, "aspect"))
plot(hs , col= grey((0:255)/255), legend = FALSE)
}
\references{
Karkee, M., Steward, B., Aziz, S., 2008. Improving quality of
  public domain digital elevation models through data fusion. Biosyst. Eng.
  101, 293-305. doi:10.1016/j.biosystemseng.2008.09.010
}
