% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_trend_surface.R
\name{fit_trend_surface}
\alias{fit_trend_surface}
\title{Fits a trend surface by least-squares to a RasterLayer.}
\usage{
fit_trend_surface(x, size = ncell(x), np = 3)
}
\arguments{
\item{x}{\code{RasterObject}.}

\item{size}{positive integer giving the number of items to choose}

\item{np}{
degree of polynomial surface
}

\item{sampleProportion}{numeric. Fraction. Approximate proportion of pixels
that will be passed on to \code{\link[spatial]{surf.ls}} as xyz coordinates.
\code{\link[raster]{sampleRandom}} is used with the argument \code{size} as
\code{ncell(x) * sampleProportion}.}
}
\value{
\code{RasterObject}.
}
\description{
Fits a trend surface by least-squares to a RasterLayer.
}
\details{
Fits a trend surface by least-squares to an object of class
  RasterLayer. This function is powered by \code{\link[spatial]{surf.ls}}.
}
\examples{
r <- raster()
r[] <- seq(1, ncell(r)) + rnorm(ncell(r), 20000, 10000)
plot(r)
s <- fit_trend_surface(r, np = 1)
plot(s)
par(mfrow = c(1,2))
plot(r, legend = FALSE)
plot(r - s, legend = FALSE)
}
