% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_over_tiles.R
\name{iterate_over_tiles}
\alias{iterate_over_tiles}
\title{Apply a function in tiles}
\usage{
iterate_over_tiles(x, fun, tile_size = floor(sqrt(blockSize(x)$nrows[1] *
  ncol(x))), overlapping_percentage = NULL, show_progress_bar = TRUE,
  filename = "", ...)
}
\arguments{
\item{x}{\code{\linkS4class{Raster}}}

\item{fun}{function}

\item{tile_size}{numeric. Tile size in pixels.}

\item{filename}{character. To create a file during the process.}

\item{...}{Additional arguments as for \code{\link{writeRaster}}}
}
\value{
\code{\linkS4class{Raster}}
}
\description{
Apply a function in tiles.
}
\details{
Useful if you are running out of \emph{RAM} while running spatial sensitive
  process.
}
\examples{
/inst/examples/iterate_over_tiles_example.R
}
