% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_artifacts.R
\name{mask_artifacts}
\alias{mask_artifacts}
\title{Mask artifacts}
\usage{
mask_artifacts(x, kernel_sizes = 5, times_sd = 3)
}
\arguments{
\item{x}{\code{linkS4class{RasterLayer}}}

\item{kernel_sizes}{numeric. Size of the kernels, could be one or more. See details.}

\item{times_sd}{numeric. Thresold value. See details.}
}
\value{
x \code{linkS4class{RasterLayer}}. Binary image.
}
\description{
Mask artifacts
}
\details{
Is a faster approach that I pick up from Hengl and Reuter (2009).
  The algorithm compute the average elevation with a kernel and compute the
  difference between the average elevation and the elevation at the core
  cell. A threshold is used to classify this difference.

  If users uses several kernels, the masks are aggregated.
}
\examples{
r <- raster(ncol = 200, nrow = 100)
projection(r) <- NA
extent(r) <- extent(0, 200, 0, 100)

set.seed(11)
dem <- fake_dem(r, n_random_data = ncell(r) * 0.01, z_range = c(0, 500))

mask_artifacts(m)

plot(m)
}
\references{
Hengl, T., Reuter, H.I., 2009. GEOMORPHOMETRY Concepts, Software,
  Applications, Developmen. ed. Elsevier, Amsterdam. (p. 100)
}
