% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_raster.R
\name{smooth_dem}
\alias{smooth_dem}
\title{Smooth Raster Layer}
\usage{
smooth_dem(x, theta)
}
\arguments{
\item{x}{\code{\linkS4class{RasterLayer}}. Single Layer.}

\item{...}{Arguments passed on to \code{fields::image.smooth}
\describe{
  \item{theta}{the bandwidth or scale parameter.}
}}
}
\value{
\code{\linkS4class{RasterLayer}}
}
\description{
Wrapper function for \code{\link[fields]{image.smooth}}.
}
\examples{
r <- raster()
r[] <- seq(1, ncell(r)) + rnorm(ncell(r), 20000, 10000)
plot(r)
r.smoothed <- smooth_dem(r, 10)
plot(r.smoothed)


}
